Create lotr_rpg.py
A simple yet fun little text-based RPG based on Lord of the Rings and characters from the series. You get to choose your character’s race which determines your fighting abilities and the rest of your company. Move through the story to defeat various enemy characters and fight the dragon to claim the treasure.